Break out matcher/compare.hpp
