+++ include/trompeloeil.hpp
@@ -27,6 +27,7 @@
 // * If a macro kills a kitten, this threatens extinction of all felines!
 
 #include "trompeloeil/mock.hpp"
+#include "trompeloeil/matcher/compare.hpp"
 #include "trompeloeil/matcher/re.hpp"
 
 #endif // include guard
